style(namespace): organiza arquivos usando namespace
